Move home assets to (home) folder
