+++ app/(home)/_components/AboutMe/index.module.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/mixins';
+@use '../../../../styles/mixins';
 
 #intro_text {
     margin: 0.7em 0.3em 0;
