+++ app/(home)/_components/Ratings/index.module.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/mixins';
+@use '../../../../styles/mixins';
 
 .rating_list {
     font-weight: bold;
