+++ app/(home)/_components/Ratings/index.tsx
