+++ app/(home)/_components/TempTwitter/index.tsx
@@ -2,9 +2,9 @@ import microCMS from "@/lib/microCMS";
 import matter from "gray-matter";
 import React from "react";
 import {Suspense} from "react";
-import LoadingBlock from "../../LoadingBlock";
+import LoadingBlock from "../../../../components/LoadingBlock";
 import ReactMarkdown from "react-markdown";
-import MainWrapper from "../../MainWrapper";
+import MainWrapper from "../../../../components/MainWrapper";
 import Block from "@/components/Block";
 import remarkGfm from "remark-gfm";
 import rehypeRaw from "rehype-raw";
