+++ app/(home)/_components/TopPageBalloons.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import Block from "@/components/Block";
-import styles from "@/app/style.module.scss";
+import styles from "@/app/(home)/style.module.scss";
 import Link from "next/link";
 import { useBalloonState } from "@/app/balloon/BalloonArray";
 import Balloon from "@/app/balloon/Balloon";
