+++ app/(home)/_components/TopPageIcons.tsx
@@ -1,5 +1,5 @@
 import Block from "@/components/Block";
-import styles from "@/app/style.module.scss";
+import styles from "@/app/(home)/style.module.scss";
 import Image from "next/legacy/image";
 import Link from "next/link";
