+++ app/(home)/_components/WhatsNew/index.module.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/mixins';
+@use '../../../../styles/mixins';
 
 #whats_new_table {
     overflow-y: scroll;
