+++ app/(home)/style.module.scss
@@ -1,4 +1,4 @@
-@use '../styles/top-page/grid';
+@use '../../styles/top-page/grid';
 
 .main-window {
     margin: 1em;
